requirements.txt
openai
llama_index
llama-index-embeddings-huggingface
fastapi 
uvicorn
python-dotenv
torch 
sentence-transformers
